refactor: Rename ramda by use and rework runner.js
